chore(textlint-fixer-formatter): move textlint/test/fixer -> textlint-fixer-formatter/test
