+++ packages/textlint-fixer-formatter/test/fixtures/double-fixed.md
