+++ packages/textlint-fixer-formatter/test/fixtures/double.js
