+++ packages/textlint-fixer-formatter/test/fixtures/double.md
