+++ packages/textlint-fixer-formatter/test/fixtures/multiple-fixed.md
