+++ packages/textlint-fixer-formatter/test/fixtures/multiple.js
