+++ packages/textlint-fixer-formatter/test/fixtures/remaining.js
