+++ packages/textlint-fixer-formatter/test/fixtures/single-fixed.md
