+++ packages/textlint-fixer-formatter/test/fixtures/single.js
