+++ packages/textlint-fixer-formatter/test/ts-setup.js
@@ -1,0 +1,3 @@
+require("ts-node").register({
+    project: __dirname
+});
